Implement permission to CRUD server variables (#521) (#584)

+++ src/variable/variable.controller.ts
@@ -2,7 +2,7 @@ import { Body, Controller, Delete, Get, NotFoundException, Param, Patch, Post, R
 import { ApiSecurity } from '@nestjs/swagger';
 import { VariableService } from 'variable/variable.service';
 import { PolyAuthGuard } from 'auth/poly-auth-guard.service';
-import { CreateVariableDto, Role, UpdateVariableDto, VariableDto, Visibility } from '@poly/model';
+import { CreateVariableDto, Permission, UpdateVariableDto, VariableDto, Visibility } from '@poly/model';
 import { AuthData, AuthRequest } from 'common/types';
 import { AuthService } from 'auth/auth.service';
 
@@ -15,17 +15,24 @@ export class VariableController {
   ) {
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Get()
   async getVariables(@Req() req: AuthRequest): Promise<VariableDto[]> {
     const environmentId = req.user.environment.id;
+
+    await this.authService.checkPermissions(req.user, [
+      Permission.ManageSecretVariables,
+      Permission.ManageNonSecretVariables,
+      Permission.Use,
+    ]);
+
     const variables = await this.service.getAll(environmentId);
     return Promise.all(
       variables.map(async variable => await this.service.toDto(variable)),
     );
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Post()
   async createVariable(@Req() req: AuthRequest, @Body() data: CreateVariableDto): Promise<VariableDto> {
     const {
@@ -36,15 +43,23 @@ export class VariableController {
       visibility = Visibility.Environment,
       secret = false,
     } = data;
+
+    await this.authService.checkPermissions(req.user, secret
+      ? Permission.ManageSecretVariables
+      : Permission.ManageNonSecretVariables);
+
     const variable = await this.service.createVariable(req.user.environment.id, context, name, description, value, visibility, secret);
 
     return this.service.toDto(variable);
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Get(':id')
   async getVariable(@Req() req: AuthRequest, @Param('id') id: string): Promise<VariableDto> {
     const variable = await this.findVariable(req.user, id);
+
+    await this.authService.checkPermissions(req.user, Permission.Use);
+
     return this.service.toDto(variable);
   }
 
@@ -55,10 +70,13 @@ export class VariableController {
     if (variable.secret) {
       throw new NotFoundException(`Variable with id '${id}' is secret`);
     }
+
+    await this.authService.checkPermissions(req.user, Permission.Use);
+
     return this.service.getVariableValue(variable);
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Patch(':id')
   async updateVariable(@Req() req: AuthRequest, @Param('id') id: string, @Body() data: UpdateVariableDto): Promise<VariableDto> {
     const {
@@ -71,15 +89,29 @@ export class VariableController {
     } = data;
     const variable = await this.findVariable(req.user, id);
 
+    await this.authService.checkPermissions(
+      req.user,
+      secret ?? variable.secret
+        ? Permission.ManageSecretVariables
+        : Permission.ManageNonSecretVariables,
+    );
+
     return this.service.toDto(
       await this.service.updateVariable(req.user.environment.id, variable, name, context, description, value, visibility, secret),
     );
   }
 
-  @UseGuards(new PolyAuthGuard([Role.Admin]))
+  @UseGuards(PolyAuthGuard)
   @Delete(':id')
   async deleteVariable(@Req() req: AuthRequest, @Param('id') id: string): Promise<void> {
     const variable = await this.findVariable(req.user, id);
+
+    await this.authService.checkPermissions(
+      req.user,
+      variable.secret
+        ? Permission.ManageSecretVariables
+        : Permission.ManageNonSecretVariables,
+    );
 
     await this.service.deleteVariable(variable);
   }
